requirements.txt
Django>=3.2
djangorestframework>=3.12
pillow>=8.0
pandas>=1.1
django-filter>=2.4
django-cors-headers>=3.7
openpyxl>=3.0
